Fixed issue with integer overflow in AVR bootloader CAN module.
Former-commit-id: 86dab1132ec8594aae1f17beec268a0a1b4ef856 [formerly b3df9f717511677bcba0e8bf7f85b21eecbb8da5] [formerly 63ed7b46bfe9dd1ee36024d7099769287fb852cb [formerly 1b4b5fe5532bd5866b68aada4b2812c34ab13634]]
Former-commit-id: 7cda73fda678884757027c09f83229c8b302a05f [formerly 5600acbac4dc00ad3384dc036915623634bcf762]
Former-commit-id: 966de38a991434d848fa52c5710ea7723c01bb4a

+++ res/avr/bootloader/bootloader_module_can.cpp
@@ -120,10 +120,10 @@
 // DECLARE PRIVATE GLOBAL VARIABLES.
 
 // Time in units of MODULE_EVENT_PERIOD (in ms) between sending out ALERT_UPLOADER messages while the bootloader is waiting for communications to begin.
-const uint8_t ALERT_UPLOADER_TIMEOUT = 50;
+const uint16_t ALERT_UPLOADER_TIMEOUT = 50;
 
 // Time in units of MODULE_EVENT_PERIOD (in ms) between the last message arriving, and the bootloader giving up and assuming it has lost comms.
-const uint8_t COMMS_TIMEOUT = 1000;
+const uint16_t COMMS_TIMEOUT = 1000;
 
 const uint8_t NODE_ID = <<<TC_INSERTS_NODE_ID_HERE>>>;
